+++ Lista1/integracao-1/nbproject/private/private.properties
@@ -1,0 +1,7 @@
+application.args=C:\\Users\\affon\\Desktop\\bbb.txt
+compile.on.save=true
+do.depend=false
+do.jar=true
+javac.debug=true
+javadoc.preview=true
+user.properties.file=C:\\Users\\affon\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
