+++ README.md
@@ -1,1 +1,13 @@
 # i-2018
+
+- Você pode colocar alguns detalhes aqui, conforme está, i-2018 é uma incógnita!
+
+- Também forneça detalhes que permitam te identificar por meio da sua presença no github.
+
+- Renomeei um dos diretórios usando o que é esperado em um projeto de software.
+Acentos não são recomendados.
+
+- Seu projeto depende do NetBeans, nada contra, nem a favor. Para o código ter
+independência do IDE, é preciso usar uma ferramenta de build. Em Java, minha 
+sugestão, derivada "do mundo", é usar o Maven (http://maven.apache.org/).
+
